Make progress towards transitioning to SHA-256.
* Remove generic hash interface and instead create a thin wrapper over the
SHA-256 implementation.
* Modify bootrom to use SHA-256 (building).
* Modfiy sm to use SHA-256 (building).
* Create additional symlinks to the crypto directory for the sm. Consider
reworking these, as we're including `crypto/include` and `crypto`, which
is sloppy.

+++ crypto/include/crypto/string.h
@@ -3,7 +3,7 @@
 
 /* TODO: Use actual implementation of these functions. */
 
-void* memcpy(void* dest, const void* src, size_t len)
+void* memcpy(void* dest, const void* src, unsigned int len)
 {
   const char* s = src;
   char *d = dest;
@@ -14,7 +14,7 @@ void* memcpy(void* dest, const void* src, size_t len)
   return d;
 }
 
-void* memset(void* dest, int byte, size_t len)
+void* memset(void* dest, int byte, unsigned int len)
 {
   char* d = dest;
   char* end = dest + len;
